refactor http enabled clients
* create an application specific HTTPClient that can be injected into
other HTTP capable services.
* update api client so that it only supports calling out to the exercism
API; deleting any unused good.
* move debug pkg into api since it is only used there.
* inject application specific HTTPClient into HTTP enabled services
